Configure TCP or UNIX socket
A new variable `CLAMAV_SOCKET_TYPE` has been added. It can take two
values: "tcp" or "file".

On CentOS and Ubuntu 14 it is configured in the `clamd.conf` file.

On Ubuntu 16 this configuration is more complex because the
clamav-daemon package uses three systemd files to configure the socket
type (unix file or tcp socket). This PR uses debconf-set-selections to
configure the socket type, but the clamav-daemon package doesn't read
the debconf system when upgrading or reinstalling, it is used only when
installing from scratch, so a package purge is needed when changing the
socket type.

The SELinux policies `antivirus_can_scan_system` and `clamd_use_jit` will
be activated if SELinux is enabled on CentOS.

On Ubuntu task files, the "Enable services" task has been modified to
restart the clamav-daemon service instead of starting. It is needed to
apply the changes made by the role.

On Ubuntu the default socket type is UNIX file.
On Centos the default socket type is TCP.